fix: build errors — KiouMatchMode declaration order + NSInteger format
Two compile errors from the last commit:

1. kKiouMatchModeAdapterOffsets[] / kKiouMatchModeTags[] were defined
   in Internal.h BEFORE the KiouMatchMode enum was declared. Moved
   them to after the enum so the designators resolve.

2. NSInteger %d format mismatch in Hook_SettingsUI.m's "N of 5"
   caption builder. Cast KIOU_MMODE_COUNT to long with %ld.

make jailed and make binpatch both build cleanly now (verified
locally). The jailed dylib has no libsubstrate / libdobby external
dependency as expected.

+++ Sources/KiouForge/Hook_SettingsUI.m
@@ -164,7 +164,8 @@ static NSString *const kAboutTwitterURL = @"https://x.com/tkgling";
                         if (kiou_kifuModeEnabled((KiouMatchMode)i)) on++;
                     }
                     cell.detailTextLabel.text =
-                        [NSString stringWithFormat:@"%d of %d", on, KIOU_MMODE_COUNT];
+                        [NSString stringWithFormat:@"%d of %ld",
+                         on, (long)KIOU_MMODE_COUNT];
                 }
             } else {
                 cell.detailTextLabel.text =
